Relocate AnswerCompanyWorkInvite prsntr and cntrlr
