+++ arbeitszeit_flask/views/company_work_invite_view.py
@@ -4,20 +4,24 @@ from uuid import UUID
 
 from flask import Response, redirect
 
-from arbeitszeit.use_cases.answer_company_work_invite import AnswerCompanyWorkInvite
+from arbeitszeit.use_cases.answer_company_work_invite import (
+    AnswerCompanyWorkInvite,
+    AnswerCompanyWorkInviteRequest,
+)
 from arbeitszeit.use_cases.show_company_work_invite_details import (
     ShowCompanyWorkInviteDetailsUseCase,
 )
 from arbeitszeit_flask.database import commit_changes
 from arbeitszeit_flask.template import TemplateIndex, TemplateRenderer
-from arbeitszeit_web.answer_company_work_invite import (
+from arbeitszeit_web.controllers.answer_company_work_invite_controller import (
     AnswerCompanyWorkInviteController,
     AnswerCompanyWorkInviteForm,
-    AnswerCompanyWorkInvitePresenter,
-    AnswerCompanyWorkInviteRequest,
 )
 from arbeitszeit_web.controllers.show_company_work_invite_details_controller import (
     ShowCompanyWorkInviteDetailsController,
+)
+from arbeitszeit_web.presenters.answer_company_work_invite_presenter import (
+    AnswerCompanyWorkInvitePresenter,
 )
 from arbeitszeit_web.presenters.show_company_work_invite_details_presenter import (
     ShowCompanyWorkInviteDetailsPresenter,
