+++ arbeitszeit_web/presenters/answer_company_work_invite_presenter.py
@@ -1,38 +1,11 @@
 from dataclasses import dataclass
-from typing import Protocol, Union
-from uuid import UUID
 
 from arbeitszeit.use_cases.answer_company_work_invite import (
-    AnswerCompanyWorkInviteRequest,
     AnswerCompanyWorkInviteResponse,
 )
-
-from .malformed_input_data import MalformedInputData
-from .notification import Notifier
-from .session import Session
-from .translator import Translator
-from .url_index import UrlIndex
-
-
-class AnswerCompanyWorkInviteForm(Protocol):
-    def get_is_accepted_field(self) -> bool:
-        ...
-
-
-@dataclass(frozen=True)
-class AnswerCompanyWorkInviteController:
-    session: Session
-
-    def import_form_data(
-        self, form: AnswerCompanyWorkInviteForm, invite_id: UUID
-    ) -> Union[None, MalformedInputData, AnswerCompanyWorkInviteRequest]:
-        requesting_user = self.session.get_current_user()
-        if requesting_user is not None:
-            return AnswerCompanyWorkInviteRequest(
-                form.get_is_accepted_field(), invite_id, requesting_user
-            )
-        else:
-            return None
+from arbeitszeit_web.notification import Notifier
+from arbeitszeit_web.translator import Translator
+from arbeitszeit_web.url_index import UrlIndex
 
 
 @dataclass
